* 'master' of https://github.com/hansewu/MovieMator: (21 commits)
  WINDOWS日志提示WARNING：QFSFileEngine::open: No file name specified #294
  Eliminate “extern” warning #389
  The window title does not contain a '[*]' placeholder #290
  Correction of false judgment conditions
  The window title does not contain a '[*]' placeholder #290
  Remove an error code (wyl)
  fix bug： mergeWith
  KeyFrame undo、redo setKeyFrame
  remove Warning: Z-order assignment: '' is not a valid widget. in encodedock.ui
  remove warning.
  remove warnings.
  remove some simple warnings.
  Remove warning
  Remove mvcp
  Remove templateeditordock
  Remove PlaylistDock
  Fixed player default 10% bug
  remove configurationdock
  播放窗口 播放模式和音量界面优化 #335 播放窗口， 原快进，快退按钮换成前一帧，后一帧按钮 #334
  Resolve "some audio file classification errors"
  ...

# Conflicts:
#	src/MovieMator/src/widgets/lumamixtransition.cpp
#	src/MovieMator/src/widgets/lumamixtransition.h

+++ src/MovieMator/src/qmltypes/colorwheelitem.cpp
@@ -101,8 +101,8 @@ void ColorWheelItem::setBlue(int blue)
 
 int ColorWheelItem::wheelSize() const
 {
-    qreal ws = (qreal)width() / (1.0 + 1.0 / WHEEL_SLIDER_RATIO);
-    return qMin(ws, height());
+    qreal ws = qreal(width() / (1.0 + 1.0 / WHEEL_SLIDER_RATIO));
+    return int(qMin(ws, height()));
 }
 
 QColor ColorWheelItem::colorForPoint(const QPoint &point)
@@ -190,10 +190,10 @@ void ColorWheelItem::paint(QPainter *painter)
 {
     Q_ASSERT(painter);
 
-    QSize size( width(), height() );
+    QSize size( (int(width())), int(height()) );
 
     if( m_size != size ) {
-        m_image = QImage(QSize(width(), height()), QImage::Format_ARGB32_Premultiplied);
+        m_image = QImage(QSize(int(width()), int(height())), QImage::Format_ARGB32_Premultiplied);
         m_image.fill(qRgba(0,0,0,0));
         drawWheel();
         drawSlider();
@@ -260,7 +260,7 @@ void ColorWheelItem::drawSliderBar(QPainter &painter)
 {
     qreal value = 1.0 - m_color.valueF();
     int ws = wheelSize();
-    int w = (qreal)ws / WHEEL_SLIDER_RATIO;
+    int w = int(qreal(ws) / WHEEL_SLIDER_RATIO);
     int h = ws - m_margin * 2;
     QPen pen(Qt::white);
     pen.setWidth(2);
@@ -276,7 +276,7 @@ void ColorWheelItem::drawSlider()
     QPainter painter(&m_image);
     painter.setRenderHint(QPainter::Antialiasing);
     int ws = wheelSize();
-    int w = (qreal)ws / WHEEL_SLIDER_RATIO;
+    int w = int(qreal(ws) / WHEEL_SLIDER_RATIO);
     int h = ws - m_margin * 2;
     QLinearGradient gradient(0, 0, w, h);
     gradient.setColorAt(0.0, Qt::white);
